+++ tests/ImageGenerators/SvgTest.php
@@ -2,8 +2,8 @@
 
 namespace Spatie\MediaLibrary\Test\ImageGenerators;
 
-use Spatie\MediaLibrary\ImageGenerators\FileTypes\Svg;
 use Spatie\MediaLibrary\Test\TestCase;
+use Spatie\MediaLibrary\ImageGenerators\FileTypes\Svg;
 
 class SvgTest extends TestCase
 {