+++ tests/TestCase.php
@@ -3,9 +3,9 @@
 namespace Spatie\MediaLibrary\Test;
 
 use File;
-use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Database\Schema\Blueprint;
 use Orchestra\Testbench\TestCase as Orchestra;
+use Illuminate\Database\Eloquent\Relations\Relation;
 
 abstract class TestCase extends Orchestra
 {
@@ -70,8 +70,8 @@ abstract class TestCase extends Orchestra
         ]);
 
         $app['config']->set('filesystems.disks.media', [
-           'driver' => 'local',
-           'root' => $this->getMediaDirectory(),
+            'driver' => 'local',
+            'root' => $this->getMediaDirectory(),
         ]);
 
         $app['config']->set('filesystems.disks.secondMediaDisk', [
@@ -85,6 +85,7 @@ abstract class TestCase extends Orchestra
 
         $app['config']->set('app.key', '6rE9Nz59bGRbeMATftriyQjrpF7DcOQm');
 
+        $this->setupS3($app);
         $this->setUpMorphMap();
     }
 
@@ -159,5 +160,24 @@ abstract class TestCase extends Orchestra
         Relation::morphMap([
             'test-model-with-morph-map' => TestModelWithMorphMap::class,
         ]);
+    }
+
+    private function setupS3($app)
+    {
+        $s3Configuration = [
+            'driver' => 's3',
+            'key' => getenv('S3_ACCESS_KEY_ID'),
+            'secret' => getenv('S3_SECRET_ACCESS_KEY'),
+            'region' => getenv('S3_BUCKET_REGION'),
+            'bucket' => getenv('S3_BUCKET_NAME'),
+        ];
+
+        $this->canTestS3 = ! (bool) array_search(false, $s3Configuration);
+
+        $app['config']->set('filesystems.disks.s3', $s3Configuration);
+        $app['config']->set(
+            'laravel-medialibrary.s3.domain',
+            'https://'.$s3Configuration['bucket'].'.s3.amazonaws.com'
+        );
     }
 }